Created View's for all screens, added Annual Yield calculations to Stocks, condensed some data in stocks, minor UI changes to Portfolio.
