+++ PortfolioHub/app/(tabs)/calendar.tsx
@@ -1,9 +1,9 @@
 import React, { useState, useCallback } from 'react';
 import { useFocusEffect } from '@react-navigation/native';
 import useCalendarViewModel from '@/viewmodels/CalendarViewModel';
-import CalendarView from '@/components/CalendarView';
+import CalendarView from '@/view/CalendarView';
 
-const ThemedCalendar = () => {
+const Calendar = () => {
   const { fetchCalendarEvents } = useCalendarViewModel();
   const [loading, setLoading] = useState<boolean>(false);
 
@@ -31,4 +31,4 @@ const ThemedCalendar = () => {
   );
 };
 
-export default ThemedCalendar;
+export default Calendar;
