+++ PortfolioHub/app/(tabs)/home.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
-import HomeView from '@/components/HomeView';
+import HomeView from '@/view/HomeView';
 
-const HomePage = () => {
+const Home = () => {
   return (
       <HomeView/>
   );
 };
 
-export default HomePage;
+export default Home;
