+++ PortfolioHub/app/ResetPassword.tsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import ResetPasswordView from '@/view/ResetPasswordView';
+
+const ResetPassword = () => {
+  return (
+      <ResetPasswordView/>
+  );
+};
+
+export default ResetPassword;
