+++ PortfolioHub/app/index.tsx
@@ -1,73 +1,10 @@
-import { View, Text, StyleSheet, TouchableOpacity } from 'react-native';
-import { ResizeMode, Video } from 'expo-av';
-import { Link } from 'expo-router';
-import { defaultStyles } from '@/constants/Styles';
-import useIntroViewModel from '@/viewmodels/IntroViewModel';
+import React from 'react';
+import IntroView from '@/view/IntroView';
 
 const Intro = () => {
-  const { videoUri, headerText } = useIntroViewModel();
-
   return (
-    <View style={styles.container}>
-      {videoUri && (
-        <Video
-          resizeMode={ResizeMode.COVER}
-          isMuted
-          isLooping
-          shouldPlay
-          source={{ uri: videoUri }}
-          style={styles.video}
-        />
-      )}
-      <View style={{ marginTop: 80, padding: 20 }}>
-        <Text style={styles.header}>{headerText}</Text>
-      </View>
-
-      <View style={styles.buttons}>
-        <Link
-          href={'/login'}
-          style={[defaultStyles.pillButton, { flex: 1, backgroundColor: '#141518' }]}
-          asChild>
-          <TouchableOpacity>
-            <Text style={{ color: 'white', fontSize: 22, fontWeight: '500' }}>Login</Text>
-          </TouchableOpacity>
-        </Link>
-        <Link
-          href={'/signup'}
-          style={[defaultStyles.pillButton, { flex: 1, backgroundColor: '#fff' }]}
-          asChild>
-          <TouchableOpacity>
-            <Text style={{ fontSize: 22, fontWeight: '500' }}>Sign Up</Text>
-          </TouchableOpacity>
-        </Link>
-      </View>
-    </View>
+      <IntroView/>
   );
 };
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    justifyContent: 'space-between',
-  },
-  video: {
-    width: '100%',
-    height: '100%',
-    position: 'absolute',
-  },
-  header: {
-    fontSize: 36,
-    fontWeight: '900',
-    textTransform: 'uppercase',
-    color: 'white',
-  },
-  buttons: {
-    flexDirection: 'row',
-    justifyContent: 'center',
-    gap: 20,
-    marginBottom: 60,
-    paddingHorizontal: 20,
-  },
-});
 
 export default Intro;
